+++ vite.config.ts
@@ -1,7 +1,76 @@
-import { defineConfig } from "vite";
+import { defineConfig, loadEnv, ConfigEnv, UserConfig } from "vite";
 import vue from "@vitejs/plugin-vue";
+import { resolve } from "path";
+import { wrapperEnv } from "./src/utils/getEnv";
+import { visualizer } from "rollup-plugin-visualizer";
+import viteCompression from "vite-plugin-compression";
+import vueJsx from "@vitejs/plugin-vue-jsx";
 
-// https://vitejs.dev/config/
-export default defineConfig({
-	plugins: [vue()]
+// @see: https://vitejs.dev/config/
+export default defineConfig(({ mode }: ConfigEnv): UserConfig => {
+	const env = loadEnv(mode, process.cwd());
+	const viteEnv = wrapperEnv(env);
+
+	return {
+		resolve: {
+			alias: {
+				"@": resolve(__dirname, "./src"),
+				"#": resolve(__dirname, "./src/types")
+			}
+		},
+		server: {
+			// 服务器主机名，如果允许外部访问，可设置为 "0.0.0.0"
+			host: "0.0.0.0",
+			port: viteEnv.VITE_PORT,
+			open: viteEnv.VITE_OPEN,
+			cors: true,
+			proxy: {
+				"/api": {
+					target: "", // easymock
+					changeOrigin: true,
+					rewrite: path => path.replace(/^\/api/, "/api/v1")
+				}
+			}
+		},
+		plugins: [
+			vue(),
+			// vite 可以使用 jsx/tsx 语法
+			vueJsx(),
+			// 是否生成包预览(分析依赖包大小,方便做优化处理)
+			viteEnv.VITE_REPORT && visualizer(),
+			// gzip compress
+			viteCompression({
+				verbose: true,
+				disable: viteEnv.VITE_BUILD_GZIP,
+				threshold: 10240,
+				algorithm: "gzip",
+				ext: ".gz"
+			})
+		],
+		// 打包去除 console.log && debugger
+		esbuild: {
+			pure: viteEnv.VITE_DROP_CONSOLE ? ["console.log", "debugger"] : []
+		},
+		build: {
+			outDir: "dist",
+			minify: "esbuild",
+			// esbuild 打包更快，但是不能去除 console.log，terser打包慢，但能去除 console.log
+			// minify: "terser",
+			// terserOptions: {
+			// 	compress: {
+			// 		drop_console: viteEnv.VITE_DROP_CONSOLE,
+			// 		drop_debugger: true
+			// 	}
+			// },
+			chunkSizeWarningLimit: 1500,
+			rollupOptions: {
+				output: {
+					// Static resource classification and packaging
+					chunkFileNames: "assets/js/[name]-[hash].js",
+					entryFileNames: "assets/js/[name]-[hash].js",
+					assetFileNames: "assets/[ext]/[name]-[hash].[ext]"
+				}
+			}
+		}
+	};
 });