+++ src/locales/modules/en.ts
@@ -1,11 +1,49 @@
 export default {
-	menus: {
-		fourZeroThree: 403,
-		fourZeroFour: 404,
-		fiveZeroZero: 500
+	router: {
+		403: "Page 403",
+		404: "Page 404",
+		500: "Page 500"
 	},
-	button: {
-		login: "login",
-		logout: "logout"
+	global: {
+		title: "Smile Admin",
+		logout: {
+			button: "Logout",
+			confirm: "Do you confirm to exit the system?",
+			success: "Logout successfully"
+		},
+		loginForm: {
+			label: {
+				account: "Account or Email",
+				password: "Password"
+			},
+			placeholder: {
+				account: "Please Input Account or Email",
+				password: "Please Input Password"
+			},
+			schema: {
+				account: "Please Input Account or Email",
+				password: "The password must be at least six characters long"
+			},
+			button: {
+				login: "Login",
+				forget: "Forget Password",
+				qq: "QQ",
+				wechat: "WeChat",
+				github: "Github"
+			},
+			text: {
+				title: "Login",
+				remember: "Remember Me",
+				OAuth: "OAuth Login"
+			}
+		},
+		welcome: {
+			hello_message: "Hello,",
+			hello_morning_01: "Good morning",
+			hello_morning_02: "Good morning",
+			hello_afternoon_01: "Good afternoon",
+			hello_afternoon_02: "Good afternoon",
+			hello_evening: "Good evening"
+		}
 	}
 };