+++ src/locales/modules/zh.ts
@@ -1,11 +1,49 @@
 export default {
-	menus: {
-		fourZeroThree: 403,
-		fourZeroFour: 404,
-		fiveZeroZero: 500
+	router: {
+		403: "403页面",
+		404: "404页面",
+		500: "500页面"
 	},
-	button: {
-		login: "登录",
-		logout: "退出登录"
+	global: {
+		title: "Smile 后台管理系统",
+		logout: {
+			button: "退出登录",
+			confirm: "是否确认退出系统？",
+			success: "退出成功"
+		},
+		loginForm: {
+			label: {
+				account: "邮箱或手机号",
+				password: "密码"
+			},
+			placeholder: {
+				account: "请输入账号或邮箱",
+				password: "请输入密码"
+			},
+			schema: {
+				account: "请输入邮箱或手机号",
+				password: "密码不能少于6位"
+			},
+			button: {
+				login: "登录",
+				forget: "忘记密码",
+				qq: "QQ",
+				wechat: "微信",
+				github: "Github"
+			},
+			text: {
+				title: "登录",
+				remember: "记住我",
+				OAuth: "第三方登录"
+			}
+		},
+		welcome: {
+			hello_message: "你好，",
+			hello_morning_01: "早上好",
+			hello_morning_02: "上午好",
+			hello_afternoon_01: "中午好",
+			hello_afternoon_02: "下午好",
+			hello_evening: "晚上好"
+		}
 	}
 };