+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - 新增 eslint、prettier、stylelint、husky 配置
 - 新增 commitlint、cz-git、editorConfig 配置 和 项目编码规范说明文件
+- 新增 打包、服务、别名、打包和压缩组件等 vite 配置 和 学习记录文档等